+++ lib/sanbase_web/graphql/schema/github_types.ex
@@ -1,0 +1,8 @@
+defmodule SanbaseWeb.Graphql.GithubTypes do
+  use Absinthe.Schema.Notation
+
+  object :activity_point do
+    field(:datetime, non_null(:datetime))
+    field(:activity, non_null(:integer))
+  end
+end