+++ app/CHANGELOG.md
@@ -1,5 +1,16 @@
 ## CHANGELOG
 
+- 2017-29-12 | v.0.2.8:
+
+  Added storybook (dev exp.)
+  Change chart hover behavior
+
+- 2017-29-12 | v.0.2.7:
+
+  Added charts with Volume, Price, Marketcap
+  Added time filter
+  Some fixes, styles
+
 - 2017-26-12 | v.0.2.6:
 
   Added Project Detailed Page